requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
psycopg2==2.9.6
PyJWT==2.7.0
pyodbc==4.0.39
python-dateutil==2.8.2
six==1.16.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
